KVM: x86/MMU: pKVM: Warn if root_to_sp() is called
With pKVM, mmu->root.hpa does not represent an actual physical address
of a shadow page (as there are no shadow pages) and thus calling
root_to_sp() is incorrect. Code paths that call it should not be
triggered if pKVM is enabled. Add a WARN_ON_ONCE so that if it is still
triggered for some reason, we'll notice that.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>