pKVM: x86: Add helper to handle the PV interfaces for a loaded vCPU
The subsequent patches will add new PV interfaces for a loaded guest
vCPU. For such PV interfaces, group them in a dedicated function so that
the corresponding pkvm_vcpu can be obtained from the per-CPU pointer
cur_guest_vcpu rather than relying on the input from the host.

vCPU switching between the guest and the host are performed around the
PV interface handling.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>