pKVM: x86: Redefine KVM print macros for the pKVM hypervisor
The KVM print macros may be included in the code which are leveraged
by the pKVM hypervisor. Overridden these print macros to use the "pkvm"
prefixes, and avoid printing current task information which is not
supported in the pKVM hypervisor.

Disable CONFIG_DYNAMIC_DEBUG and CONFIG_DYNAMIC_DEBUG_CORE to strip off
the unsupported print symbols for pKVM non-debug version.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>