pKVM: x86: Import kvm_set_msr_common() from KVM
Import kvm_set_msr_common() from KVM for the pKVM. This will be used to
emulate MSR writing.

A MSR emulation policy is defined to guarantee the pVM's MSR security.
If a MSR is not supported by the kvm_set_msr_common(), whether it can be
emulated by the host is depending on this policy:
  - The TSC_ADJUST and TSC are not secure to pVM. Current pVM use cases
    treat TSC as untrusted and reply on different source of secure time.
  - The x2apic related MSRs can be emulated by the host as virtual APIC
    is emulated by the host.
  - All other emulated MSRs can only be emulated by the host for an npVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>