KVM: pVMX: Implement segment related operations
Enable the pKVM VMX host to access guest segments via the segment PV
interfaces. Skipping the segment PV interfaces once the guest state is
marked as protected as the future accessing guest segment is no longer
permitted by the pKVM hypervisor.

The host maintains a cache mask to indicate whether a segment is already
read form the pKVM or not, and only use the PV interface to read if it
is not. As the cache mask will only be updated after getting the guest
segment from the pKVM via the PV interface, the cache mask can correctly
reflect the segment state regardless the preemption, no need to provide
a no_cache version.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>