pKVM: x86: Add event injection PV interfaces
Add event injection PV interfaces for the host to inject irq, nmi and
exceptions via vendor-specific operations.

The PV interfaces for injecting irq/nmi are always accessible to the
host as the host emulates the irq/nmi injections for both npVM and pVM.
The exception for pVM is emulated by the pKVM, thus making this PV
interface always inaccessible to the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>