KVM: pVMX: Implement handle_exit operation
Enable the pKVM VMX host to handle the various vmexit reasons. The
vcpu_run failures will be handled to similarly with KVM VMX host, by
setting the detailed failure information in the vcpu_run page and return
to the user space VMM.

The detailed vmexit reasons will be handled by the handlers in the
pkvm_vmx_exit_handlers[], which will be added by the subsequent patches.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>