pKVM: VMX: Handle EXIT_REASON_IO_INSTRUCTION
The EXIT_REASON_IO_INSTRUCTION is mostly handled by the host as the IO
devices are emulated by the host alone. However the string IO
instruction needs to be decoded by the host, which cannot be supported
for the pVM due to isolation. So the pVM should be enlightened to unroll
the string IO instructions. In case this happened, inject #GP to the pVM.

Once the IO instruction is handled, this IO instruction should be
skipped. But due to isolation, the host can only skip instruction for
the npVM but not for the pVM. So the pKVM hypervisor will skip this
instruction before the next vmenter.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>