pKVM: VMX: Sync vCPU state for EXIT_REASON_MSR_READ
Before returning back to the host with the EXIT_REASON_MSR_READ, the
vCPU RCX and exit instruction length will be shared with the host. And
Before entering the guest again, update the vCPU RAX/RDX with the host
emulated result and skip the rdmsr instruction.

Specifically, all the npVM's vCPU registers are synced unconditionally
thus only need to share the exit instruction length before returning
back to the host so that the host can skip the rdmsr instruction. For
the pVM, no need to share the exit instruction length as the rdmsr
instruction is skipped by the pKVM hypervisor rather than the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>