pKVM: VMX: Handle EXIT_REASON_VMCALL
For the npVM, the EXIT_REASON_VMCALL is handled alone by the host. The
exit instruction length is shared with the host to skip the vmcall
instruction.

For the pVM, the EXIT_REASON_VMCALL is handled by the pKVM hypervisor.
Currently no vmcall is supported thus just simply return -KVM_EPERM to
the pVM. And #GP will be injected if the vmcall is send from user mode
to make sure vmcall can only be used in the kernel mode.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>