pKVM: VMX: Handle EXIT_REASON_EPT_MISCONFIG
For the pVM, there should not cause EPT_MISCONFIG exit as the pVM should
be enlightened to use PV method to access MMIO rather than the memory
mapped way. In case this is happened, inject #GP to the pVM.

For the npVM, returns back to the host to handle the EPT_MISCONFIG by
emulating the instruction.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>