pKVM: x86: Request host to reset guest MMU for EFER emulation
If guest is changing the EFER page table related bits, i.e., EFER_LME
and EFER_NX, the guest mmu instance should be reset according to the
new EFER value from the host side to make sure the guest mmu state is
up-to-date for instruction emulation. As the host can only emulate
instruction for npVM, set the HOST_RESET_MMU request for the npVM only.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>