pKVM: x86: Introduce pkvm_vm_handle
Introduce pkvm_vm_handle to identify a guest VM instance created by
the pKVM hypervisor. The handle is allocated during VM initialization
from a global pkvm_vms_bitmap, which is protected by a lock to ensure
safe concurrent access. And the handle is returned to the host via the
vm_init PV interface.

For simplicity, pkvm_vms_bitmap is statically defined to support up to
64 guest VMs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>