pKVM: VMX: Implement vcpu_put operation
Import the vmx_vcpu_put() for the pKVM to implement vcpu_put operation.
Unlike KVM, pKVM lacks an SMP call mechanism to notify remote CPUs to
clear VMCS when a vCPU is migrated. Therefore, VMCS clearing is performed
immediately during the vcpu_put operation.

A subsequent patch will introduce a paravirtualized vcpu_put interface to
allow the host to defer vCPU cleanup until it is scheduled on another CPU.
This avoids unnecessary VMCS load/put cycles when the vCPU remains on the
same CPU.

Note: The vcpu_put operation does not handle posted interrupts, as
interrupt management is delegated to the host. Additionally, preparing to
switch to host is required when the vCPU exits its run loop immediately
but not here when putting a vCPU.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>