pKVM: VMX: Implement set_virtual_apic_mode operation
Import vmx_set_virtual_apic_mode() from KVM to implement
set_virtual_apic_mode operation for pKVM. As emulating the virtual xapic
mode in the host requires instruction decoding, this is not doable to
host as pVM's memory and cpu state are isolated. So not to support xapic
mode for pVMs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>