pKVM: VMX: Implement pkvm_x86_ops
Add initial implementation for the pkvm_x86_ops to update and share the
pKVM owned vCPU state.

The ops update_vcpu_state_from_host will be used to update the pKVM
owned vCPU state according to different vmexit reasons. Detailed
implementation will be added by the subsequent patches.

The ops share_vcpu_state_with_host is implemented to share the vmexit
information with the host, e.g., vmexit reason, exit qualification and
so on. Share the npVM's AR field of CS/SS for optimization purposes as
these are more frequently used by the host, namely: by
get_cs_db_l_bits to emulate instructions, and by get_cpl_no_cache to
schedule out a non-protected vCPU at runtime, so that the host doesn't
need to issue additional hypercalls to retrieve the AR field of CS/SS
for that. No need to share for a pVM as the host does not emulate
instructions or use CPL for scheduling in case of a pVM anyway.

More vCPU state will be shared according to the detailed vmexit reason
which will be implemented by the subsequent patches.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>