pKVM: VMX: Implement pkvm_x86_ops
Add initial implementation for the pkvm_x86_ops to update and share the
pKVM owned vCPU state.

The ops update_vcpu_state_from_host will be used to update the pKVM
owned vCPU state according to different vmexit reasons. Detailed
implementation will be added by the subsequent patches.

The ops share_vcpu_state_with_host is implemented to share the vmexit
information with the host, e.g., vmexit reason, exit qualification and
so on. For npVM, CS/SS segment is also shared unconditionally with the
host as these may be needed for instruction emulation. No need to share
if is pVM as the host cannot emulate instruction for pVM. More vCPU
state will be shared according to the detailed vmexit reason, which will
be implemented by the subsequent patches.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>