pKVM: VMX: Introduce prepare_switch_to_host
Import vmx_prepare_switch_to_host from KVM and use it as the new
operation prepare_switch_to_host in the kvm_x86_ops for the pKVM
hypervisor, to pair with the existing prepare_switch_to_guest operation.
As the tss and fixmap of gdt are not used in the pKVM hypervisor, skip
the related operations.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>