pKVM: x86: Audit PV interfaces for protected VMs
The pVM's state should be isolated by the pKVM hypervisor from the host.
A PV interface should be audited before allowing it to run for a pVM.
The policy is that, as the host needs to pre-configure the pVM's vCPU
state for booting, the protection for pVM is only enforced by the pKVM
hypervisor once the vCPU has started running.

There is no isolation between non-protected VMs and the host, thus it is
not necessary to audit any PV interfaces for an npVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>