pKVM: x86: Add interrupt shadow PV interfaces
Add interrupt shadow PV interfaces for the host to access guest
interrupt shadow via the vendor-specific operations. As the interrupt
shadow is used by the host user space VMM to set the initial event via
ioctl, or by the host KVM to emulate certain instructions, none of these
are needed for a pVM. So these PV interfaces don't allow the host to
access the pVM's interrupt shadow.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>