pKVM: VMX: Handle EXIT_REASON_TASK_SWITCH
The EXIT_REASON_TASK_SWITCH vmexit is handled differently for the pVM
and npVM. For the pVM, injecting #GP as the pKVM hypervisor doesn't
support instruction emulation and the host cannot emulate task switch.
For the npVM, the vmexit is handled cooperatively by the pKVM hypervisor
and the host. The pKVM clears the event injection if the task switch is
caused by the task gate event and skips the instruction if the task
switch is caused by instructions. Then returns back to the host to
emulate the task switch.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>